refactor(xcm-utils): move HoldingCredit/asset_to_holding to sibling module
The trait-glue needed to feed WeightTrader::buy_weight in the fee-estimation
precompile is mostly type-system boilerplate, not precompile logic. Pull it
out into `holding.rs` so lib.rs stays focused on the precompile surface.
